+++ packages/app2/package.json
@@ -5,7 +5,8 @@
         "dev": "vinxi dev",
         "dev:tsc": "tsc -w",
         "build": "vinxi build",
-        "start": "vinxi start"
+        "start": "vinxi start",
+        "start:env": "dotenv -- vinxi start"
     },
     "dependencies": {
         "@auth/core": "^0.31.0",
@@ -59,6 +60,7 @@
         "node": ">=18"
     },
     "devDependencies": {
+        "dotenv-cli": "^7.3.0",
         "drizzle-kit": "^0.19.13",
         "postcss-import": "^16.1.0"
     }
