+++ packages/viz/src/loader/run-files-loader.ts
@@ -1,13 +1,13 @@
 import { combineRecordings, parseRecordingFile } from '@hkviz/parser';
 import { createDeferred, createMemo, createSignal } from 'solid-js';
 import { storeInitializer } from 'src/store';
-import { fetchWithRunfileCache } from './recording-file-browser-cache';
+import { fetchWithRunfileCache, openRunfileCache } from './recording-file-browser-cache';
 import { type RunFileInfo } from './run-files-info';
 import { wrapResultWithProgress } from './wrap-result-with-progress';
 import { isServer } from 'solid-js/web';
 
-async function loadFile(file: RunFileInfo, onProgress: (progress: number) => void) {
-    const loader = () => fetchWithRunfileCache(file.id, file.version, file.signedUrl);
+async function loadFile(cache: Promise<Cache | null>, file: RunFileInfo, onProgress: (progress: number) => void) {
+    const loader = () => fetchWithRunfileCache(cache, file.id, file.version, file.signedUrl);
 
     // uncomment to get file content from console. useful for debugging
     // very bad for performance, since context needs to be kept in memory.
@@ -46,11 +46,12 @@ export function createRunFileLoader(files: RunFileInfo[]): RunFileLoader {
 
     const abortController = new AbortController();
 
+    const cache = openRunfileCache();
     const fileLoaders = files.map((file) => {
         const [progress, setProgress] = createSignal(0);
         return {
             progress,
-            promise: loadFile(file, setProgress),
+            promise: loadFile(cache, file, setProgress),
         };
     });
